refactor(issue-147): extract gedcom export logic into src/lib/gedcom-export.ts
Move queries, row-mapping helpers, groupByUnionId, buildFamRecord, and a
new buildGedcomDocument (extracted from the inline main() body) into a
shared module. scripts/export-gedcom.ts is now a thin CLI wrapper that
re-exports the moved symbols and only handles Neo4j session setup and
writing the output file.
